% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{calculate_sleep}
\alias{calculate_sleep}
\title{Analysis Functions for RMB Package}
\usage{
calculate_sleep(object, sleep_threshold = 5, bin_minutes = 30)
}
\arguments{
\item{object}{MonitorS4 object}

\item{sleep_threshold}{Minimum minutes of inactivity to be considered sleep (default: 5)}

\item{bin_minutes}{Bin size in minutes for sleep analysis (default: 30)}
}
\value{
MonitorS4 object with sleep assay added
}
\description{
Functions for analyzing Drosophila behavioral data
Calculate sleep analysis from activity data
}
