% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-esoph.R
\docType{data}
\name{esoph}
\alias{esoph}
\title{Ille-et-Vilaine study data}
\format{
A data frame with 975 rows and 7 variables:
\describe{
\item{case}{Case status (1=case, 0=control)}
\item{age}{Age or age category at baseline (as coded in source data).}
\item{agegp}{Age-group category.}
\item{tob}{Tobacco consumption gm/day}
\item{tobgp}{Grouped tobacco consum.}
\item{alc}{Alcohol consumption gm/day}
\item{alcgp}{Grouped alcohol consum.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/esoph.dta}{UCSF companion data}
}
\usage{
esoph
}
\description{
Dataset used in Chapter 5 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Unmatched case-control study of esophageal cancer and alcohol/tobacco exposures in France.
Primary article: \href{https://doi.org/10.1002/ijc.2910200210}{Tuyns and colleagues (1977), \emph{International Journal of Cancer}}.
}
\keyword{datasets}
