% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fecfat.R
\docType{data}
\name{fecfat}
\alias{fecfat}
\title{Fecal fat data}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{fecfat}{Fecal fat (g/day)}
\item{subject}{Patient ID number}
\item{pilltype}{Type of pill}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/fecfat.dta
}
\usage{
fecfat
}
\description{
Dataset used in Chapter 7 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\keyword{datasets}
