% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fit.R
\docType{data}
\name{fit}
\alias{fit}
\title{FIT data}
\format{
A data frame with 5324 rows and 9 variables:
\describe{
\item{treat}{Treatment assignment}
\item{age}{Calculated baseline age}
\item{smoking}{Baseline smoking status}
\item{frac_base}{History of vertebral fracture}
\item{bmd_base}{Baseline femoral neck BMD}
\item{newvfx}{New vertebral fracture}
\item{fitpy}{Fit follow-up time (years)}
\item{bmd_pctchg}{Percent change in BMD}
\item{bmd_diff}{Change in BMD}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/fit.dta}{UCSF companion data}
}
\usage{
fit
}
\description{
Dataset used in Chapter 6 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Randomized Fracture Intervention Trial of alendronate in postmenopausal women.
Primary article: \href{https://doi.org/10.1016/S0140-6736(96)07088-2}{Black and colleagues (1996), \emph{The Lancet}}.
}
\keyword{datasets}
