% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fitdata.R
\docType{data}
\name{fitdata}
\alias{fitdata}
\title{FIT BMD data}
\format{
A data frame with 5813 rows and 10 variables:
\describe{
\item{tx}{Treatment assignment}
\item{age}{Calculated baseline age}
\item{bmi}{Baseline BMI (kg/m^2)}
\item{smoking}{Baseline smoking status}
\item{blnspfx}{History of non-spine fracture}
\item{blvfx}{History of vertebral fracture}
\item{blbmd}{Baseline femoral neck BMD}
\item{cobmd}{End of study femoral neck BMD}
\item{newvfx}{New vertebral fracture}
\item{newnspfx}{New non-spine fracture}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/fitdata.dta}{UCSF companion data}
}
\usage{
fitdata
}
\description{
Dataset used in Chapter 9 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Longitudinal FIT subset focused on bone mineral density outcomes.
Primary article: \href{https://doi.org/10.1016/S0140-6736(96)07088-2}{Black and colleagues (1996), \emph{The Lancet}}.
}
\keyword{datasets}
