% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{fitglm}
\alias{fitglm}
\title{Fracture risk data}
\format{
A data frame with 6459 rows and 22 variables:
\describe{
\item{alloc}{ALLOCATION NUMBER}
\item{ra_age}{AGE}
\item{frx}{ANY FRX}
\item{nosp}{ANY FRX NOT SPINE}
\item{numnosp}{NUMBER OF NON-SPINE FRX}
\item{trt01}{FIT TREATMENT 1=ALN 0=PLB}
\item{p3_weigh}{Weight}
\item{htotbmd}{Total Hip BMD}
\item{nbmd}{Neck BMD}
\item{trialyrs}{YEARS TO DOD OR LAST CONTACT (NEW_LAST)}
\item{frxdays}{DAYS TO FIRST FRX}
\item{nospdays}{DAYS TO FIRST NON-SPINE FRACTURE}
\item{riskcat4}{FALL RISK: 2 OR MORE FALLS, TUAG 12.5 SECONDS}
\item{tneck}{FEM NECK TSCORE, NH98}
\item{bmd25}{BMD FEMORAL NECK T SCORE <= -2.5}
\item{frxyrs}{YEARS TO FIRST FRX}
\item{nospyrs}{YEARS TO FIRST NON-SPINE FRACTURE}
\item{hplac}{HIGH FALL RISK, PLACEBO GROUP}
\item{htrt}{HIGH FALL RISK, TREATMENT GROUP}
\item{lplac}{LOW FALL RISK, PLACEBO GROUP}
\item{ltrt}{LOW FALL RISK, TREATMENT GROUP}
\item{rtgroup}{FOUR LEVEL RISK/TREATMENT GROUP}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/FITglm.dta
}
\usage{
fitglm
}
\description{
Dataset used in Chapter 8 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\keyword{datasets}
