% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gababies.R
\docType{data}
\name{gababies}
\alias{gababies}
\title{Georgia birthweight data}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
\item{momid}{Mother identifier for clustered births.}
\item{birthord}{Birth order for the infant within each mother.}
\item{momage}{Maternal age at delivery (years).}
\item{timesnc}{Time since previous birth (interpregnancy interval).}
\item{lowbrth}{Indicator for low birthweight infant.}
\item{bweight}{Infant birthweight (grams).}
\item{delwght}{Maternal weight at delivery.}
\item{lastwght}{Most recent maternal weight before delivery.}
\item{initage}{Mom's age at first birth}
\item{initwght}{Maternal initial/prepregnancy weight.}
\item{cinitage}{Centered maternal age at initial pregnancy.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/gababies.dta}{UCSF companion data}
}
\usage{
gababies
}
\description{
Dataset used in Chapter 7 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Clustered repeated-birth sample from Georgia mothers used for birthweight analyses.
Primary article: \href{https://regression.ucsf.edu/second-edition/data-examples-and-problems}{RMB2e repeated-birth Georgia cohort example}.
}
\keyword{datasets}
