% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_metadata_template}
\alias{generate_metadata_template}
\title{Generate metadata template based on monitor data}
\usage{
generate_metadata_template(raw_data, monitor_name)
}
\arguments{
\item{raw_data}{Parsed monitor data}

\item{monitor_name}{Name of the monitor}
}
\value{
A data.frame with metadata template
}
\description{
Generate metadata template based on monitor data
}
