% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{getAssay}
\alias{getAssay}
\title{Get assay data from MonitorS4 object}
\usage{
getAssay(object, assay = NULL)
}
\arguments{
\item{object}{A MonitorS4 object}

\item{assay}{Character string specifying which assay to retrieve}
}
\value{
The specified assay data
}
\description{
Get assay data from MonitorS4 object
}
