% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{getMeta}
\alias{getMeta}
\title{Get metadata from MonitorS4 object}
\usage{
getMeta(object)
}
\arguments{
\item{object}{A MonitorS4 object}
}
\value{
The metadata data.frame
}
\description{
Get metadata from MonitorS4 object
}
