% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-leuk.R
\docType{data}
\name{leuk}
\alias{leuk}
\title{Leukemia data}
\format{
A data frame with 42 rows and 3 variables:
\describe{
\item{time}{weeks in remission}
\item{cens}{relapse indicator}
\item{group}{treatment group}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/leuk.dta}{UCSF companion data}
}
\usage{
leuk
}
\description{
Dataset used in Chapter 3 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Randomized acute lymphoblastic leukemia remission trial comparing 6-MP versus placebo.
}
\keyword{datasets}
