% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_files}
\alias{match_files}
\title{Match monitor files with metadata files}
\usage{
match_files(data_path, metadata_path)
}
\arguments{
\item{data_path}{Path to directory containing Monitor files}

\item{metadata_path}{Path to directory containing metadata files}
}
\value{
A data.frame matching monitor files to metadata files
}
\description{
Match monitor files with metadata files
}
