% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRMB.R
\name{merge_monitors}
\alias{merge_monitors}
\title{Merge multiple MonitorS4 objects}
\usage{
merge_monitors(monitor_list)
}
\arguments{
\item{monitor_list}{List of MonitorS4 objects to merge}
}
\value{
Single merged MonitorS4 object
}
\description{
Merge multiple MonitorS4 objects
}
