% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{MonitorS4}
\alias{MonitorS4}
\title{Constructor for MonitorS4 objects}
\usage{
MonitorS4(
  meta.data = data.frame(),
  assays = list(),
  active.assay = character(0),
  time = list()
)
}
\arguments{
\item{meta.data}{A data.frame with metadata (optional)}

\item{assays}{A list of assay data (optional)}

\item{active.assay}{Character string for active assay (optional)}

\item{time}{A list with time data (optional)}
}
\value{
A MonitorS4 object
}
\description{
Constructor for MonitorS4 objects
}
