% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mros.R
\docType{data}
\name{mros}
\alias{mros}
\title{MrOS data}
\format{
A data frame with 5994 rows and 5 variables:
\describe{
\item{bmd3}{3 quantiles of bmd}
\item{months}{Follow-up time in months.}
\item{status}{Event status indicator (coded outcome/censoring status).}
\item{weight}{q2.weight (kg)}
\item{years}{Follow-up time in years.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/mros.dta}{UCSF companion data}
}
\usage{
mros
}
\description{
Dataset used in Chapter 6 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Prospective cohort of older men in the Osteoporotic Fractures in Men (MrOS) study.
Primary article: \href{https://doi.org/10.1016/j.cct.2005.05.006}{Orwoll and colleagues (2005), \emph{Contemporary Clinical Trials}}.
}
\keyword{datasets}
