% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_monitor_file}
\alias{parse_monitor_file}
\title{Utility Functions for RMB Package}
\usage{
parse_monitor_file(file_path)
}
\arguments{
\item{file_path}{Path to Monitor*.txt file}
}
\value{
A data.frame with parsed monitor data
}
\description{
Various utility functions for data processing and manipulation
Parse TriKinetics monitor data file
}
