% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pbc.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{PBC data}
\format{
A data frame with 312 rows and 23 variables:
\describe{
\item{number}{Case Number}
\item{status}{Alive/Dead}
\item{rx}{treatment}
\item{sex}{sex}
\item{asictes}{Presence of Asictes}
\item{hepatom}{Hepatomegaly}
\item{spiders}{Presence of Spiders}
\item{edema}{Presence of Edema}
\item{bilirubin}{Serum Bilirubin in mg/dl}
\item{cholest}{Serum Cholesterol in mg/dl}
\item{albumin}{Albumin in gm/dl}
\item{copper}{Urine Copper in ug/day}
\item{alkphos}{Alkaline Phosphatase in U/liter}
\item{sgot}{SGOT in U/ml}
\item{trigli}{Triglicerides in mg/dl}
\item{platel}{Platelets per cubic ml / 1000}
\item{prothrom}{Prothrombin time in seconds}
\item{histol}{Histologic stage of disease}
\item{age}{Age (years)}
\item{years}{Time to Death (in Years)}
\item{logbili}{Bilirubin (log)}
\item{logalbu}{Albumin (log)}
\item{logprot}{Prothom time (log)}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/pbc.dta
}
\usage{
pbc
}
\description{
Dataset used in Chapter 6 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\keyword{datasets}
