% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_monitor_assays}
\alias{process_monitor_assays}
\title{Process monitor data into different assay types}
\usage{
process_monitor_assays(raw_data)
}
\arguments{
\item{raw_data}{Parsed monitor data from parse_monitor_file}
}
\value{
A list with different assay types (mt, ct, pn)
}
\description{
Process monitor data into different assay types
}
