% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_dead_flies}
\alias{remove_dead_flies}
\title{Remove dead flies from MonitorS4 object}
\usage{
remove_dead_flies(object, activity_threshold = 1, consecutive_zeros = 12)
}
\arguments{
\item{object}{MonitorS4 object}

\item{activity_threshold}{Minimum activity threshold to consider alive}

\item{consecutive_zeros}{Maximum consecutive zeros allowed}
}
\value{
MonitorS4 object with dead flies removed
}
\description{
Remove dead flies from MonitorS4 object
}
