% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmb-package.R
\docType{package}
\name{rmb-package}
\alias{rmb}
\alias{rmb-package}
\title{rmb: Datasets for Regression Methods in Biostatistics}
\description{
Provides datasets from the companion website for "Regression Methods in Biostatistics" (2nd edition), including variable labels from the source files and vignettes with exploratory analyses and example models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/rmb}
  \item Report bugs at \url{https://github.com/d-morrison/rmb/issues}
}

}
\author{
\strong{Maintainer}: Ezra Morrison \email{demorrison@ucdavis.edu}

}
