% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMB.R
\name{RMB}
\alias{RMB}
\title{Robust Median-Based Bayesian Model Wrapper}
\usage{
RMB(Missing_Type, data, seed, chain = 1, Niter = 6000, burnIn = 3000)
}
\arguments{
\item{Missing_Type}{Type of missing mechanism ("MAR", "MNAR", "MCAR", or "no missing")}

\item{data}{Input data matrix}

\item{seed}{Random seed}

\item{chain}{Number of MCMC chains}

\item{Niter}{Total number of MCMC iterations}

\item{burnIn}{Number of burn-in iterations}
}
\value{
An mcmc.list object invisibly
}
\description{
Robust Median-Based Bayesian Model Wrapper
}
