% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{rmb_datasets}
\alias{rmb_datasets}
\title{RMB Datasets Index}
\format{
A data frame with 26 rows and 4 variables:
\describe{
\item{object}{Dataset object name in this package.}
\item{title}{Dataset title from the UCSF companion website.}
\item{chapter}{Chapter where the dataset is first introduced.}
\item{url}{Original Stata download URL.}
}
}
\source{
https://regression.ucsf.edu/second-edition/data-examples-and-problems
}
\usage{
rmb_datasets
}
\description{
A catalog of datasets from \emph{Regression Methods in Biostatistics} (2nd edition).
}
\keyword{datasets}
