% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{setAssay}
\alias{setAssay}
\title{Set assay data in MonitorS4 object}
\usage{
setAssay(object, assay, data)
}
\arguments{
\item{object}{A MonitorS4 object}

\item{assay}{Character string specifying assay name}

\item{data}{Data to store in the assay}
}
\value{
Updated MonitorS4 object
}
\description{
Set assay data in MonitorS4 object
}
