% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{setMeta}
\alias{setMeta}
\title{Set metadata in MonitorS4 object}
\usage{
setMeta(object, meta.data)
}
\arguments{
\item{object}{A MonitorS4 object}

\item{meta.data}{A data.frame with metadata}
}
\value{
Updated MonitorS4 object
}
\description{
Set metadata in MonitorS4 object
}
