% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorS4-class.R
\name{show,MonitorS4-method}
\alias{show,MonitorS4-method}
\title{Show method for MonitorS4 objects}
\usage{
\S4method{show}{MonitorS4}(object)
}
\arguments{
\item{object}{A MonitorS4 object}
}
\value{
Prints object summary
}
\description{
Show method for MonitorS4 objects
}
