% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sof.R
\docType{data}
\name{sof}
\alias{sof}
\title{SOF data}
\format{
A data frame with 42950 rows and 12 variables:
\describe{
\item{id}{ID}
\item{visit}{Variable description not provided in source metadata.}
\item{totbmd}{Variable description not provided in source metadata.}
\item{nbmd}{Variable description not provided in source metadata.}
\item{vdate}{Variable description not provided in source metadata.}
\item{v2date}{Variable description not provided in source metadata.}
\item{bmi}{Variable description not provided in source metadata.}
\item{n_iadl}{Variable description not provided in source metadata.}
\item{age}{V2AGE2}
\item{meno_age}{Variable description not provided in source metadata.}
\item{del_totbmd}{Variable description not provided in source metadata.}
\item{base_totbmd}{Variable description not provided in source metadata.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/sof.dta}{UCSF companion data}
}
\usage{
sof
}
\description{
Dataset used in Chapter 7 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Prospective cohort from the Study of Osteoporotic Fractures in older women.
}
\keyword{datasets}
