% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-whickham.R
\docType{data}
\name{whickham}
\alias{whickham}
\title{Whickham data}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{vstatus}{Vital status at follow-up.}
\item{smoker}{Indicator for current smoking.}
\item{agegrp}{Age-group category.}
\item{nn}{Participant sequence number in the cohort sample.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/whickam.dta}{UCSF companion data}
}
\usage{
whickham
}
\description{
Dataset used in Chapter 3 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Population-based prospective cohort from Whickham, England, with 20-year mortality follow-up.
Primary article: \href{https://pubmed.ncbi.nlm.nih.gov/8733876/}{Vanderpump and colleagues (1996), \emph{Thyroid}}.
}
\keyword{datasets}
