% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert wide format data to long format for analysis}
\usage{
wide_to_long(data, value_name = "value")
}
\arguments{
\item{data}{Data frame in wide format}

\item{value_name}{Name for the value column}
}
\value{
Data frame in long format
}
\description{
Convert wide format data to long format for analysis
}
