% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-actg019.R
\docType{data}
\name{actg019}
\alias{actg019}
\title{ACTG 019 data}
\format{
A data frame with 880 rows and 5 variables:
\describe{
\item{id}{Subject ID #}
\item{days}{Days Since Randomization}
\item{cens}{Event Indicator}
\item{rx}{Treatment}
\item{cd4}{Baseline CD4+ lymphocyte count}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/actg019.dta}{UCSF companion data}
}
\usage{
actg019
}
\description{
Dataset used in Chapter 6 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Randomized double-blind placebo-controlled trial of zidovudine in HIV infection.
}
\keyword{datasets}
