% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fecfat.R
\docType{data}
\name{fecfat}
\alias{fecfat}
\title{Fecal fat data}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{fecfat}{Fecal fat (g/day)}
\item{subject}{Patient ID number}
\item{pilltype}{Type of pill}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/fecfat.dta}{UCSF companion data}
}
\usage{
fecfat
}
\description{
Dataset used in Chapter 7 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Repeated-measures study comparing pancreatic enzyme regimens using fecal fat outcomes.
Primary article: \href{https://regression.ucsf.edu/second-edition/data-examples-and-problems}{RMB2e chapter example based on historical fecal-fat treatment data}.
}
\keyword{datasets}
