% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figure4_1.R
\docType{data}
\name{figure4_1}
\alias{figure4_1}
\title{Figure 4.1 data}
\format{
A data frame with 8 rows and 1 variables:
\describe{
\item{x2}{Variable description not provided in source metadata.}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/figure4_1.dta
}
\usage{
figure4_1
}
\description{
Dataset used in Chapter 4 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\keyword{datasets}
