% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figure4_12.R
\docType{data}
\name{figure4_12}
\alias{figure4_12}
\title{Figure 4.12 data}
\format{
A data frame with 26 rows and 4 variables:
\describe{
\item{y1}{Variable description not provided in source metadata.}
\item{y2}{Variable description not provided in source metadata.}
\item{y3}{Variable description not provided in source metadata.}
\item{x}{Variable description not provided in source metadata.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/figure4_12.dta}{UCSF companion data}
}
\usage{
figure4_12
}
\description{
Dataset used in Chapter 4 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Illustrative teaching dataset used for model-checking examples in Chapter 4.
}
\keyword{datasets}
