% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gababies.R
\docType{data}
\name{gababies}
\alias{gababies}
\title{Georgia birthweight data}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
\item{momid}{Mother's ID}
\item{birthord}{Variable description not provided in source metadata.}
\item{momage}{Variable description not provided in source metadata.}
\item{timesnc}{Variable description not provided in source metadata.}
\item{lowbrth}{Variable description not provided in source metadata.}
\item{bweight}{Variable description not provided in source metadata.}
\item{delwght}{5 delwght}
\item{lastwght}{5 lastwght}
\item{initage}{Mom's age at first birth}
\item{initwght}{Variable description not provided in source metadata.}
\item{cinitage}{Variable description not provided in source metadata.}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/gababies.dta
}
\usage{
gababies
}
\description{
Dataset used in Chapter 7 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\details{
Study design: Clustered repeated-birth sample from Georgia mothers used for birthweight analyses.
}
\keyword{datasets}
