% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hers_nodm_longitudinal.R
\docType{data}
\name{hers_nodm_longitudinal}
\alias{hers_nodm_longitudinal}
\title{Table 11.8 HERS data}
\format{
A data frame with 9217 rows and 15 variables:
\describe{
\item{pptid}{subject id}
\item{exercise}{bl exercise &/or walk >3x/wk, 1:y 0:n}
\item{avgdrpwk}{Avg number of drinks per week}
\item{csmker}{current smoking status 1:current}
\item{weight}{Variable description not provided in source metadata.}
\item{bmi}{Body mass index}
\item{sbp}{Systolic blood pressure}
\item{glucose}{Variable description not provided in source metadata.}
\item{htnmeds}{anti-htn meds use}
\item{diabetes}{(1=yes, 0=no)}
\item{age}{age}
\item{group}{treatment assignment 0:a: pbo, 1:b:hrt}
\item{grade_hi}{highest grade completed}
\item{white}{(1=white, 0=other)}
\item{visit}{Variable description not provided in source metadata.}
}
}
\source{
\href{https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/hers_nodm_longitudinal.dta}{UCSF companion data}
}
\usage{
hers_nodm_longitudinal
}
\description{
Dataset used in Chapter 11 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
Study design: Longitudinal repeated-measures HERS subset restricted to participants without diabetes.
}
\keyword{datasets}
