% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mira_hsv.R
\docType{data}
\name{mira_hsv}
\alias{mira_hsv}
\title{MIRA data}
\format{
A data frame with 6069 rows and 6 variables:
\describe{
\item{agecat}{agecat}
\item{mos}{Variable description not provided in source metadata.}
\item{id}{Variable description not provided in source metadata.}
\item{hsv2}{Variable description not provided in source metadata.}
\item{stihx}{Variable description not provided in source metadata.}
\item{newparts}{Variable description not provided in source metadata.}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/mira_hsv.dta
}
\usage{
mira_hsv
}
\description{
Dataset used in Chapter 5 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\keyword{datasets}
