% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mros.R
\docType{data}
\name{mros}
\alias{mros}
\title{MrOS data}
\format{
A data frame with 5994 rows and 5 variables:
\describe{
\item{bmd3}{3 quantiles of bmd}
\item{months}{Variable description not provided in source metadata.}
\item{status}{Variable description not provided in source metadata.}
\item{weight}{q2.weight (kg)}
\item{years}{Variable description not provided in source metadata.}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/mros.dta
}
\usage{
mros
}
\description{
Dataset used in Chapter 6 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\details{
Study design: Prospective cohort of older men in the Osteoporotic Fractures in Men (MrOS) study.
}
\keyword{datasets}
