% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-whickham.R
\docType{data}
\name{whickham}
\alias{whickham}
\title{Whickham data}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{vstatus}{Variable description not provided in source metadata.}
\item{smoker}{Variable description not provided in source metadata.}
\item{agegrp}{1; 2; 3}
\item{nn}{Variable description not provided in source metadata.}
}
}
\source{
https://regression.ucsf.edu/sites/g/files/tkssra16191/files/wysiwyg/home/data/whickam.dta
}
\usage{
whickham
}
\description{
Dataset used in Chapter 3 of \emph{Regression Methods in Biostatistics}
(2nd edition), as distributed on the UCSF companion website.
}
\details{
Study design: Population-based prospective cohort from Whickham, England, with 20-year mortality follow-up.
}
\keyword{datasets}
